[Feat] Add interface Board-, PostRepository

+++ REST_Code/Models/Board.cs
@@ -2,14 +2,13 @@
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.Linq;
-using System.Threading.Tasks;
 
 namespace REST_Code.Models
 {
     public class Board
     {
         #region Properties
-        public int Id { get; set; }
+        public long Id { get; set; }
         [Required]
         public String Name { get; set; }
         [Required]
